Fancyyyy
Okay so I added the code that shows the list of friends and navigates to another screen when you click on a friend. The second screen displays the list of matched movies for that friend (hardcoded for now). There is also a back button to go back to the friends list