Pending Friends Class
This class is structured the same way as the userProfile (displaying liked list).
This class will take all the current pending friends of the user, create cards and display those cards in a list. User will have the option to accept or decline the pending friend. If they accept, the pending friend card should pop up in the main list of friends and disappear from pending friends.